--- api/routes/scripts/init.go
@@ -1,11 +1,0 @@
-package scripts
-
-import "os"
-
-var controlURL, isControlURLDefined = os.LookupEnv("CONTROL_API_URL")
-
-func init() {
-	if !isControlURLDefined {
-		panic("CONTROL_API_URL is required in environment")
-	}
-}
