+++ tests/control/conftest.py
@@ -1,18 +1,71 @@
 import json
 from pathlib import Path
 
+import requests
 from pytest import fixture
 
 FIXTURES_DIR = Path(__file__).resolve().parent / "fixtures"
+PATIENT_MAPPING_FILE = Path(__file__).resolve().parent.parent / "fixtures/patient_mapping.json"
 
 
-@fixture(scope="session")
+@fixture(scope="module")
 def preview_mapping():
     with open(FIXTURES_DIR / "preview_mapping.json", "r") as fp:
         return json.load(fp)
 
 
-@fixture(scope="session")
+@fixture(scope="module")
 def erroneous_mapping():
     with open(FIXTURES_DIR / "erroneous_mapping.json", "r") as fp:
         return json.load(fp)
+
+
+class MockResponse:
+    def __init__(self, json_data, status_code):
+        self.json_data = json_data
+        self.status_code = status_code
+
+    def json(self):
+        return self.json_data
+
+
+@fixture
+def mock_pyrog_response(monkeypatch):
+    def mock_post(*args, headers, **kwargs):
+        if headers["Authorization"] == "Bearer validToken":
+            with open(PATIENT_MAPPING_FILE, "r") as fp:
+                return MockResponse({"data": {"resource": json.load(fp)}}, 200)
+        elif headers["Authorization"] == "Bearer forbiddenToken":
+            return MockResponse({"errors": [{"statusCode": 403, "message": "forbidden"}]}, 200)
+        elif headers["Authorization"] == "Bearer unauthorizedToken":
+            return MockResponse({"errors": [{"statusCode": 401, "message": "unauthorized"}]}, 200)
+        else:
+            return MockResponse("invalid token", 400)
+
+    monkeypatch.setattr(requests, "post", mock_post)
+
+
+fhirConceptMap = {
+    "id": "cm_gender",
+    "group": [
+        {
+            "element": [
+                {"code": "F", "target": [{"code": "female", "equivalence": "equal"}]},
+                {"code": "M", "target": [{"code": "male", "equivalence": "equal"}]},
+            ]
+        }
+    ],
+}
+
+
+@fixture
+def mock_fhir_api_response(monkeypatch):
+    def mock_get(*args, headers):
+        if headers["Authorization"] == "Bearer validToken":
+            return MockResponse({"data": fhirConceptMap}, 200)
+        elif headers["Authorization"] == "Bearer forbiddenToken":
+            return MockResponse("invalid token", 403)
+        else:
+            return MockResponse("invalid token", 401)
+
+    monkeypatch.setattr(requests, "get", mock_get)
